only show signin required on home page
