+++ src/stores/instanceStore.ts
@@ -57,6 +57,7 @@ export const useInstanceStore = defineStore("instance", () => {
   }
 
   const isLoggedIn = computed(() => !!currentUser.value);
+  const isReady = computed(() => fetchStatus.value === "success");
 
   async function init() {
     // don't fetch if we already have data or are fetching
@@ -71,11 +72,13 @@ export const useInstanceStore = defineStore("instance", () => {
 
   return {
     fetchStatus,
+    isReady,
     pages,
     instance,
     currentUser,
     isLoggedIn,
     collections,
+
     refresh,
   };
 });
